+++ src/handlers/schedule/index.ts
@@ -1,18 +1,18 @@
 import { Client } from "discord.js";
 import schedule from "node-schedule";
-import listDir from "../../helpers/checkDirectory";
+import checkDirectory from "../../helpers/checkDirectory";
 import { IJob } from "../../interfaces/Job";
 import logger from "../../middlewares/logger";
 
 export const start = async (client: Client) => {
   logger.info("⏰ Started job management");
 
-  const jobNames = await listDir("jobs");
+  const jobNames = await checkDirectory("schedules");
   if (!jobNames) return logger.warn("No available jobs found");
 
   await Promise.all(
     jobNames.map(async (jobName) => {
-      const job: IJob = await import(`../../jobs/${jobName}`);
+      const job: IJob = await import(`../../schedules/${jobName}`);
 
       schedule.scheduleJob(job.options.schedule, async () => {
         logger.verbose(`⏰ Performed the job "${jobName}"`);
