+++ src/handlers/updatePresence/index.ts
