--- src/helpers/getPlugins/index.ts
@@ -1,10 +1,0 @@
-import logger from "../../middlewares/logger";
-import listDir from "../checkDirectory";
-
-export default async (dir: string) => {
-  logger.debug("Processing list of plugins...");
-  const plugins = await listDir(dir);
-  logger.verbose("Processed list of plugins!");
-
-  return plugins || [];
-};
