--- src/helpers/registerDiscordCommands/index.ts
@@ -1,22 +1,0 @@
-import { REST } from "@discordjs/rest";
-import { Routes, SlashCommandBuilder } from "discord.js";
-import logger from "../../middlewares/logger";
-
-export default async (builder: SlashCommandBuilder) => {
-  const rest = new REST({ version: "10" }).setToken(process.env.DISCORD_TOKEN);
-
-  await rest.put(
-    Routes.applicationGuildCommands(
-      process.env.DISCORD_CLIENT_ID,
-      process.env.DISCORD_GUILD_ID
-    ),
-    { body: builder }
-  );
-
-  await rest
-    .put(Routes.applicationCommands(process.env.DISCORD_CLIENT_ID as string), {
-      body: builder,
-    })
-    .then(() => logger.info("Successfully deployed to Discord"))
-    .catch(() => logger.error("Failed to deployed to Discord"));
-};
