+++ src/plugins/commands/manage/modules/counters/modules/add/index.ts
@@ -7,9 +7,9 @@ import {
   PermissionsBitField,
 } from "discord.js";
 // Configurations
+import prisma from "../../../../../../../handlers/database";
 import getEmbedConfig from "../../../../../../../helpers/getEmbedData";
 import logger from "../../../../../../../middlewares/logger";
-import prisma from "../../../../../../../prisma";
 
 // Function
 export default {
