+++ src/plugins/commands/reputation/modules/give/index.ts
@@ -1,11 +1,11 @@
 import { SlashCommandSubcommandBuilder } from "@discordjs/builders";
 import { ChatInputCommandInteraction, EmbedBuilder } from "discord.js";
-import { command as CooldownCommand } from "../../../../../helpers/cooldown";
+import { command as CooldownCommand } from "../../../../../handlers/cooldown";
 import getEmbedConfig from "../../../../../helpers/getEmbedData";
 import logger from "../../../../../middlewares/logger";
 import noSelfReputation from "./components/noSelfReputation";
 
-import prisma from "../../../../../prisma";
+import prisma from "../../../../../handlers/database";
 
 export default {
   metadata: { guildOnly: true, ephemeral: true },
