+++ src/plugins/events/guildCreate/index.ts
@@ -1,8 +1,8 @@
 import { Guild } from "discord.js";
-import updatePresence from "../../../helpers/updatePresence";
+import prisma from "../../../handlers/database";
+import updatePresence from "../../../handlers/updatePresence";
 import { IEventOptions } from "../../../interfaces/EventOptions";
 import logger from "../../../middlewares/logger";
-import prisma from "../../../prisma";
 
 export const options: IEventOptions = {
   type: "on",
