+++ src/plugins/events/messageCreate/modules/counters/index.ts
@@ -1,6 +1,6 @@
 import { ChannelType, Message } from "discord.js";
+import prisma from "../../../../../handlers/database";
 import logger from "../../../../../middlewares/logger";
-import prisma from "../../../../../prisma";
 
 export default {
   execute: async (message: Message) => {
