--- src/prisma.ts
@@ -1,3 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-export default new PrismaClient();
